+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/manager/CertificationIdManager.java
@@ -19,4 +19,5 @@ public interface CertificationIdManager {
 	public List<CertificationIdDTO> getAll();
 	public CertificationIdDTO create(List<Long> productIds, String year) throws EntityRetrievalException, EntityCreationException, JsonProcessingException;
 	public CertificationIdDTO create(CertificationIdDTO dto) throws EntityRetrievalException, EntityCreationException, JsonProcessingException;
+	public String getAllAsCsvString();
 }